requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
frozenlist==1.5.0
idna==3.10
iniconfig==2.0.0
load-dotenv==0.1.0
multidict==6.1.0
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
propcache==0.2.1
pydantic==2.10.3
pydantic_core==2.27.1
pytest==8.3.4
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
typing_extensions==4.12.2
tzdata==2024.2
yarl==1.18.3
